+++ server/config/db.js
@@ -1,0 +1,6 @@
+const knex = require('knex');
+const knexfile = require("../knexfile");
+
+const db =knex(knexfile.development);
+
+module.exports = db;
