+++ server/models/dataSapiModel.js
@@ -1,0 +1,11 @@
+const db = require("../config/db")
+
+class DataSapiModel{
+    getData(){
+  const data = db('datasapi') 
+  return data     
+    }
+
+}
+
+module.exports = new DataSapiModel();
