+++ server/request.rest
@@ -1,0 +1,4 @@
+get http://localhost:3000/data-sapi
+
+###
+
